Oreal-v2/requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.24.0

# Text analysis and NLP
textblob>=0.17.1
vaderSentiment>=3.3.2
langdetect>=1.0.9
emoji>=2.2.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.15.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Streamlit web app
streamlit>=1.28.0

# System monitoring (for large datasets)
psutil>=5.9.0

# NLTK data (automatically downloaded)
nltk>=3.8

# AI and LLM features
openai>=1.0.0
anthropic>=0.7.0
transformers>=4.30.0
torch>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# AI Assistant Dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0